+++ lib/src/listenable_widget.dart
@@ -1,28 +1,36 @@
-import 'dart:async';
-
 import 'package:flutter/widgets.dart';
 
 abstract class ListenableWidget<T extends ChangeNotifier>
     extends StatefulWidget {
   const ListenableWidget({super.key});
 
-  /// Used to create the [ViewModel] instance.
+  /// Creates and returns the [T] instance.
+  /// This method is called once during the widget's initialization.
   T create(BuildContext context);
 
   /// Builds the widget tree using the provided [viewModel].
+  /// This method is called whenever the [viewModel] notifies its listeners.
   Widget build(BuildContext context, T viewModel);
 
-  /// Called when the widget is updated. This can be used to refresh the
-  /// [viewModel] or perform any necessary updates.
-  FutureOr<void> update(
+  /// Called when the widget is updated with a new configuration.
+  /// This can be used to update the [viewModel] or perform any necessary actions.
+  ///
+  /// The [oldWidget] parameter contains the previous widget configuration.
+  void onWidgetChanged(
     BuildContext context,
     ListenableWidget<T> oldWidget,
     T viewModel,
-  ) async {}
+  ) {}
+
+  /// Called when the widget's dependencies change.
+  /// This can be used to refresh the [viewModel] or perform any necessary updates.
+  void onDependenciesChanged(BuildContext context, T viewModel) {}
 
   /// Whether the widget should automatically dispose of the [viewModel].
-  /// If true, the [viewModel] will be disposed when the widget is removed
-  /// from the widget tree.
+  ///
+  /// If `true` (the default), the [viewModel] will be disposed when the widget
+  /// is removed from the widget tree. Set to `false` if you manage the
+  /// [viewModel]'s lifecycle externally.
   bool get autoDispose => true;
 
   @override
@@ -42,7 +50,13 @@ class _ListenableWidgetState<T extends ChangeNotifier>
   @override
   void didUpdateWidget(covariant ListenableWidget<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    widget.update(context, oldWidget, viewModel);
+    widget.onWidgetChanged(context, oldWidget, viewModel);
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    widget.onDependenciesChanged(context, viewModel);
   }
 
   @override