+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+* refactor - update -> onWidgetChanged
+* feat - onDependencyChanged
+* refactor - Example
+* docs - Document new changes
+
 ## 0.1.0
 
 * Initial release